requirements.txt
fastapi
uvicorn
httpie
requests
httpx
jinja2
python-multipart
pillow
bcrypt
mercadopago
python-dotenv
firebase
firebase-admin
pyrebase4==4.7.1